Лекция 2.7 продолжение
Созданы классы SessionHelper, ContactHelper.
