+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/appmanager/ContactHelper.java
@@ -1,0 +1,35 @@
+package ru.stqa.pft.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import ru.stqa.pft.addressbook.model.ContactData;
+
+public class ContactHelper {
+    private WebDriver wd;
+
+    public ContactHelper(WebDriver wd) {
+        this.wd = wd;
+    }
+
+    public void returnContactHomePage() {
+        wd.findElement(By.linkText("home page")).click();
+    }
+
+    public void submitContactCreation() {
+        wd.findElement(By.xpath("//div[@id='content']/form/input[21]")).click();
+    }
+
+    public void fillContactForm(ContactData contactData) {
+        wd.findElement(By.name("firstname")).sendKeys(contactData.getFirstname());
+        wd.findElement(By.name("lastname")).click();
+        wd.findElement(By.name("lastname")).sendKeys(contactData.getLastname());
+        wd.findElement(By.name("mobile")).click();
+        wd.findElement(By.name("mobile")).sendKeys(contactData.getMobile());
+        wd.findElement(By.name("email")).click();
+        wd.findElement(By.name("email")).sendKeys(contactData.getEmail());
+    }
+
+    public void initContactCreation() {
+        wd.findElement(By.name("firstname")).click();
+    }
+}
